+++ src/libs/conduit/conduit_data_accessor.hpp
@@ -1,0 +1,157 @@
+// Copyright (c) Lawrence Livermore National Security, LLC and other Conduit
+// Project developers. See top-level LICENSE AND COPYRIGHT files for dates and
+// other details. No copyright assignment is required to contribute to Conduit.
+
+//-----------------------------------------------------------------------------
+///
+/// file: conduit_data_accessor.hpp
+///
+//-----------------------------------------------------------------------------
+
+#ifndef CONDUIT_DATA_ACCESSOR_HPP
+#define CONDUIT_DATA_ACCESSOR_HPP
+
+
+//-----------------------------------------------------------------------------
+// -- conduit  includes -- 
+//-----------------------------------------------------------------------------
+#include "conduit_node.hpp"
+
+//-----------------------------------------------------------------------------
+// -- begin conduit:: --
+//-----------------------------------------------------------------------------
+namespace conduit
+{
+
+//-----------------------------------------------------------------------------
+// -- begin conduit::DataArray --
+//-----------------------------------------------------------------------------
+///
+/// class: conduit::DataAccessor
+///
+/// description:
+///  Helps consume array data as desired type with on the fly conversion.
+///
+//-----------------------------------------------------------------------------
+template <typename T> 
+class CONDUIT_API DataAccessor
+{
+public: 
+//-----------------------------------------------------------------------------
+//
+// -- conduit::DataAccessor public methods --
+//
+//-----------------------------------------------------------------------------
+
+//-----------------------------------------------------------------------------
+// Construction and Destruction
+//-----------------------------------------------------------------------------
+        /// main constructor
+        DataAccessor(void *data, const DataType &dtype);
+        DataAccessor(Node &node);
+        ~DataAccessor();
+
+//-----------------------------------------------------------------------------
+// Data and Info Access
+//-----------------------------------------------------------------------------
+    T              operator[](index_t idx) const
+                    {return element(idx);}
+    T              element(index_t idx) const;
+
+    const void     *element_ptr(index_t idx) const
+                    {
+                         return static_cast<const char*>(m_data) +
+                                  m_dtype.element_index(idx);
+                    };
+
+    index_t         number_of_elements() const 
+                        {return m_dtype.number_of_elements();}
+
+    const DataType &dtype()    const 
+                        { return m_dtype;}
+
+private:
+
+//-----------------------------------------------------------------------------
+//
+// -- conduit::DataAccessor private data members --
+//
+//-----------------------------------------------------------------------------
+    DataAccessor();
+    /// holds data (always external, never allocated)
+    void           *m_data;
+    /// holds data description
+    DataType        m_dtype;
+    
+};
+//-----------------------------------------------------------------------------
+// -- end conduit::DataAccessor --
+//-----------------------------------------------------------------------------
+
+//-----------------------------------------------------------------------------
+//
+// -- conduit::DataAccessor typedefs for supported types --
+//
+//-----------------------------------------------------------------------------
+
+/// Note: these are also the types we explicitly instantiate.
+
+/// signed integer arrays
+typedef DataAccessor<int8>     int8_accessor;
+typedef DataAccessor<int16>    int16_accessor;
+typedef DataAccessor<int32>    int32_accessory;
+typedef DataAccessor<int64>    int64_accessor;
+
+/// unsigned integer arrays
+typedef DataAccessor<uint8>    uint8_accessor;
+typedef DataAccessor<uint16>   uint16_accessor;
+typedef DataAccessor<uint32>   uint32_accessor;
+typedef DataAccessor<uint64>   uint64_accessor;
+
+/// floating point arrays
+typedef DataAccessor<float32>  float32_accessor;
+typedef DataAccessor<float64>  float64_accessor;
+
+/// native c types arrays
+typedef DataAccessor<char>       char_accessor;
+typedef DataAccessor<short>      short_accessor;
+typedef DataAccessor<int>        int_accessor;
+typedef DataAccessor<long>       long_accessor;
+#ifdef CONDUIT_HAS_LONG_LONG
+typedef DataAccessor<long long>  long_long_accessor;
+#endif
+
+
+/// signed integer arrays
+typedef DataAccessor<signed char>       signed_char_accessor;
+typedef DataAccessor<signed short>      signed_short_accessor;
+typedef DataAccessor<signed int>        signed_int_accessor;
+typedef DataAccessor<signed long>       signed_long_accessor;
+#ifdef CONDUIT_HAS_LONG_LONG
+typedef DataAccessor<signed long long>  signed_long_long_accessor;
+#endif
+
+
+/// unsigned integer arrays
+typedef DataAccessor<unsigned char>   unsigned_char_accessor;
+typedef DataAccessor<unsigned short>  unsigned_short_accessor;
+typedef DataAccessor<unsigned int>    unsigned_int_accessor;
+typedef DataAccessor<unsigned long>   unsigned_long_accessor;
+#ifdef CONDUIT_HAS_LONG_LONG
+typedef DataArray<unsigned long long>  unsigned_long_long_accessor;
+#endif
+
+
+/// floating point arrays
+typedef DataAccessor<float>   float_accessor;
+typedef DataAccessor<double>  double_accessory;
+#ifdef CONDUIT_USE_LONG_DOUBLE
+typedef DataAccessor<long double>  long_double_accessor;
+#endif
+
+}
+//-----------------------------------------------------------------------------
+// -- end conduit:: --
+//-----------------------------------------------------------------------------
+
+#endif