+++ src/tests/conduit/t_conduit_node_type_dispatch.cpp
@@ -57,7 +57,7 @@ private:
 // -- type specific template imp 
 template<typename T>
 float64
-template_dispatch_detail(Node &n)
+pointer_dispatch_detail(Node &n)
 {
     float64 res = 0;
     index_t nele = n.dtype().number_of_elements();
@@ -71,24 +71,60 @@ template_dispatch_detail(Node &n)
 }
 
 //-----------------------------------------------------------------------------
-// -- use template dispatch
+// -- array type specific template imp 
+template<typename T>
+float64
+array_dispatch_detail(Node &n)
+{
+    float64 res = 0;
+    T t_array = n.value();
+    index_t nele = n.dtype().number_of_elements();
+    // T *t_ptr = n.value();
+    for(index_t i=0; i < nele;i++)
+    {
+        res += t_array[i];
+    }
+
+    return res;
+}
+
+//-----------------------------------------------------------------------------
+// -- use template dispatch (FASTEST but does not handle strange strides)
 float64 
-template_dispatch(Node &n)
+pointer_dispatch(Node &n)
 {
     if(n.dtype().is_float64())
     {
-        return template_dispatch_detail<float64>(n);
+        return pointer_dispatch_detail<float64>(n);
     }
     else if(n.dtype().is_float32())
     {
-        return template_dispatch_detail<float32>(n);
+        return pointer_dispatch_detail<float32>(n);
     }
 
     return 0.0;
 }
 
 //-----------------------------------------------------------------------------
-// -- convert each element to desired type in the loop
+// -- use template data array dispatch (this handles all strides!)
+float64 
+array_dispatch(Node &n)
+{
+    if(n.dtype().is_float64())
+    {
+        return array_dispatch_detail<float64_array>(n);
+    }
+    else if(n.dtype().is_float32())
+    {
+        return array_dispatch_detail<float32_array>(n);
+    }
+
+    return 0.0;
+}
+
+
+//-----------------------------------------------------------------------------
+// -- convert each element to desired type in the loop (this handles all strides)
 float64 
 on_the_fly_convert(Node &n)
 {
@@ -106,7 +142,7 @@ on_the_fly_convert(Node &n)
 }
 
 //-----------------------------------------------------------------------------
-// -- convert the array to desired type, then loop
+// -- convert the array to desired type, then loop (this handles all strides)
 float64 
 array_convert(Node &n)
 {
@@ -120,6 +156,24 @@ array_convert(Node &n)
     for(index_t i=0; i < nele;i++)
     {
         res += f64_ptr[i];
+    }
+
+    return res;
+}
+
+
+//-----------------------------------------------------------------------------
+// -- use accessor
+float64 
+data_accessor(Node &n)
+{
+    float64_accessor vals(n);
+    float64 res = 0;
+    index_t nele = vals.number_of_elements();
+
+    for(index_t i=0; i < nele;i++)
+    {
+        res += vals[i];
     }
 
     return res;
@@ -141,41 +195,57 @@ TEST(conduit_node, test_dispatch)
     std::cout << "Number of elements: " <<  ARRAY_SIZE << std::endl;
     std::cout << "Iterations: " << NUM_ITERS << std::endl;
 
-    float64 res1,res2,res3;
+    float64 res1,res2,res3,res4,res5;
 
     Timer t1;
     for(index_t i=0;i<NUM_ITERS;i++)
     {
         res1 = on_the_fly_convert(n);
     }
-    t1.elapsed();
+    float t1_tval = t1.elapsed();
 
     Timer t2;
     for(index_t i=0;i<NUM_ITERS;i++)
     {
         res2 = array_convert(n);
     }
-    t2.elapsed();
+    float t2_tval = t2.elapsed();
 
     Timer t3;
     for(index_t i=0;i<NUM_ITERS;i++)
     {
-        res3 = template_dispatch(n);
+        res3 = pointer_dispatch(n);
     }
-    t3.elapsed();
+    float t3_tval = t3.elapsed();
+
+    Timer t4;
+    for(index_t i=0;i<NUM_ITERS;i++)
+    {
+        res4 = array_dispatch(n);
+    }
+    float t4_tval = t4.elapsed();
+
+
+    Timer t5;
+    for(index_t i=0;i<NUM_ITERS;i++)
+    {
+        res5 = data_accessor(n);
+    }
+    float t5_tval = t5.elapsed();
+
 
     EXPECT_EQ(res1,float64(ARRAY_SIZE));
     EXPECT_EQ(res2,float64(ARRAY_SIZE));
     EXPECT_EQ(res3,float64(ARRAY_SIZE));
+    EXPECT_EQ(res4,float64(ARRAY_SIZE));
+    EXPECT_EQ(res5,float64(ARRAY_SIZE));
 
-    // std::cout << "on_the_fly_convert (last res): " << res1 << std::endl;
-    // std::cout << "array_convert (last res):      " << res2 << std::endl;
-    // std::cout << "template_dispatch (last res):  " << res3 << std::endl;
 
-    std::cout << "on_the_fly_convert: " << t1.elapsed() << std::endl;
-    std::cout << "array_convert:      " << t2.elapsed() << std::endl;
-    std::cout << "template_dispatch:  " << t3.elapsed() << std::endl;
-
+    std::cout << "on_the_fly_convert: " << t1_tval << std::endl;
+    std::cout << "array_convert:      " << t2_tval << std::endl;
+    std::cout << "pointer_dispatch:   " << t3_tval << std::endl;
+    std::cout << "array_dispatch:     " << t4_tval << std::endl;
+    std::cout << "data_accessor:      " << t5_tval << std::endl;
 
 }