+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 
 ## Unreleased
 
+### Added
+
+#### General
+- Added DataAccessor class that helps write generic algorithms that consume data arrays using expected types.
+
 ### Changed
 
 #### General